+++ src/canvas/StaticCanvas.ts
@@ -1,4 +1,3 @@
-import { getFabricDocument } from '../env';
 import { config } from '../config';
 import { CENTER, iMatrix, VERSION } from '../constants';
 import type { CanvasEvents, StaticCanvasEvents } from '../EventTypeDefs';
@@ -26,9 +25,8 @@ import {
   cancelAnimFrame,
   requestAnimFrame,
 } from '../util/animation/AnimationFrameProvider';
-import { getElementOffset } from '../util/dom_misc';
 import { uid } from '../util/internals/uid';
-import { createCanvasElement, isHTMLCanvas, toDataURL } from '../util/misc/dom';
+import { createCanvasElement, toDataURL } from '../util/misc/dom';
 import { invertTransform, transformPoint } from '../util/misc/matrix';
 import type { EnlivenObjectOptions } from '../util/misc/objectEnlive';
 import {
@@ -44,6 +42,8 @@ import {
   isPattern,
   isTextObject,
 } from '../util/typeAssertions';
+import { StaticCanvasDOMManager } from './DOMManagers/StaticCanvasDOMManager';
+import type { CSSDimensions } from './DOMManagers/util';
 
 export type TCanvasSizeOptions = {
   backstoreOnly?: boolean;
@@ -243,9 +243,13 @@ export class StaticCanvas<
    * Can be use to read the raw pixels, but never write or manipulate
    * @type HTMLCanvasElement
    */
-  declare lowerCanvasEl: HTMLCanvasElement;
+  get lowerCanvasEl() {
+    return this.elements.lower?.el;
+  }
 
-  declare contextContainer: CanvasRenderingContext2D;
+  get contextContainer() {
+    return this.elements.lower?.ctx;
+  }
 
   /**
    * Width in virtual/logical pixels of the canvas.
@@ -275,16 +279,11 @@ export class StaticCanvas<
    */
   declare disposed?: boolean;
 
-  /**
-   * Keeps a copy of the canvas style before setting retina scaling and other potions
-   * in order to return it to original state on dispose
-   * @type string
-   */
-  declare _originalCanvasStyle?: string;
-
   declare _offset: { left: number; top: number };
   protected declare hasLostContext: boolean;
   protected declare nextRenderHandle: number;
+
+  declare elements: StaticCanvasDOMManager;
 
   static ownDefaults: Record<string, any> = StaticCanvasDefaults;
 
@@ -307,16 +306,15 @@ export class StaticCanvas<
     this.set(options);
     this.initElements(el);
     this._setDimensionsImpl({
-      width: this.width || this.lowerCanvasEl.width || 0,
-      height: this.height || this.lowerCanvasEl.height || 0,
+      width: this.width || this.elements.lower.el.width || 0,
+      height: this.height || this.elements.lower.el.height || 0,
     });
     this.viewportTransform = [...this.viewportTransform];
     this.calcViewportBoundaries();
   }
 
   protected initElements(el: string | HTMLCanvasElement) {
-    this._createLowerCanvas(el);
-    this._originalCanvasStyle = this.lowerCanvasEl.style.cssText;
+    this.elements = new StaticCanvasDOMManager(el);
   }
 
   add(...objects: FabricObject[]) {
@@ -378,52 +376,12 @@ export class StaticCanvas<
     return this._isRetinaScaling() ? Math.max(1, config.devicePixelRatio) : 1;
   }
 
-  protected _initRetinaScaling() {
-    this.__initRetinaScaling(this.lowerCanvasEl, this.contextContainer);
-  }
-
-  protected __initRetinaScaling(
-    canvas: HTMLCanvasElement,
-    context: CanvasRenderingContext2D
-  ) {
-    const scaleRatio = config.devicePixelRatio;
-    canvas.setAttribute('width', (this.width * scaleRatio).toString());
-    canvas.setAttribute('height', (this.height * scaleRatio).toString());
-    context.scale(scaleRatio, scaleRatio);
-  }
-
   /**
    * Calculates canvas element offset relative to the document
    * This method is also attached as "resize" event handler of window
    */
   calcOffset() {
-    return (this._offset = getElementOffset(this.lowerCanvasEl));
-  }
-
-  /**
-   * Creates a bottom canvas
-   * @private
-   * @param {HTMLElement} [canvasEl]
-   */
-  protected _createLowerCanvas(canvasEl: HTMLCanvasElement | string) {
-    // canvasEl === 'HTMLCanvasElement' does not work on jsdom/node
-    if (isHTMLCanvas(canvasEl)) {
-      this.lowerCanvasEl = canvasEl;
-    } else {
-      this.lowerCanvasEl =
-        (getFabricDocument().getElementById(canvasEl) as HTMLCanvasElement) ||
-        createCanvasElement();
-    }
-    if (this.lowerCanvasEl.hasAttribute('data-fabric')) {
-      /* _DEV_MODE_START_ */
-      throw new Error(
-        'fabric.js: trying to initialize a canvas that has already been initialized'
-      );
-      /* _DEV_MODE_END_ */
-    }
-    this.lowerCanvasEl.classList.add('lower-canvas');
-    this.lowerCanvasEl.setAttribute('data-fabric', 'main');
-    this.contextContainer = this.lowerCanvasEl.getContext('2d')!;
+    return (this._offset = this.elements.calcOffset());
   }
 
   /**
@@ -471,24 +429,24 @@ export class StaticCanvas<
    * @protected
    */
   protected _setDimensionsImpl(
-    dimensions: Partial<TSize>,
+    dimensions: Partial<TSize | CSSDimensions>,
     { cssOnly = false, backstoreOnly = false }: TCanvasSizeOptions = {}
   ) {
-    Object.entries(dimensions).forEach(([prop, value]) => {
-      let cssValue = `${value}`;
+    if (!cssOnly) {
+      const size = {
+        width: this.width,
+        height: this.height,
+        ...(dimensions as Partial<TSize>),
+      };
+      this.elements.setDimensions(size, this.getRetinaScaling());
+      this.hasLostContext = true;
+      this.width = size.width;
+      this.height = size.height;
+    }
+    if (!backstoreOnly) {
+      this.elements.setCSSDimensions(dimensions);
+    }
 
-      if (!cssOnly) {
-        this._setBackstoreDimension(prop as keyof TSize, value);
-        cssValue += 'px';
-        this.hasLostContext = true;
-      }
-
-      if (!backstoreOnly) {
-        this._setCssDimension(prop as keyof TSize, cssValue);
-      }
-    });
-
-    this._isRetinaScaling() && this._initRetinaScaling();
     this.calcOffset();
   }
 
@@ -512,29 +470,6 @@ export class StaticCanvas<
     if (!cssOnly) {
       this.requestRenderAll();
     }
-  }
-
-  /**
-   * Helper for setting width/height
-   * @private
-   * @param {String} prop property (width|height)
-   * @param {Number} value value to set property to
-   * @todo subclass in canvas and handle upperCanvasEl there.
-   */
-  _setBackstoreDimension(prop: keyof TSize, value: number) {
-    this.lowerCanvasEl[prop] = value;
-    this[prop] = value;
-  }
-
-  /**
-   * Helper for setting css width/height
-   * @private
-   * @param {String} prop property (width|height)
-   * @param {String} value value to set property to
-   * @todo subclass in canvas and handle upperCanvasEl there.
-   */
-  _setCssDimension(prop: keyof TSize, value: string) {
-    this.lowerCanvasEl.style[prop] = value;
   }
 
   /**
@@ -627,7 +562,7 @@ export class StaticCanvas<
    * @return {HTMLCanvasElement}
    */
   getElement(): HTMLCanvasElement {
-    return this.lowerCanvasEl;
+    return this.elements.lower.el;
   }
 
   /**
@@ -643,7 +578,7 @@ export class StaticCanvas<
    * @return {CanvasRenderingContext2D}
    */
   getContext(): CanvasRenderingContext2D {
-    return this.contextContainer;
+    return this.elements.lower.ctx;
   }
 
   /**
@@ -655,7 +590,7 @@ export class StaticCanvas<
     this.overlayImage = null;
     this.backgroundColor = '';
     this.overlayColor = '';
-    this.clearContext(this.contextContainer);
+    this.clearContext(this.getContext());
     this.fire('canvas:cleared');
     this.renderOnAddRemove && this.requestRenderAll();
   }
@@ -668,7 +603,7 @@ export class StaticCanvas<
     if (this.destroyed) {
       return;
     }
-    this.renderCanvas(this.contextContainer, this._objects);
+    this.renderCanvas(this.getContext(), this._objects);
   }
 
   /**
@@ -1634,7 +1569,8 @@ export class StaticCanvas<
    * @throws if aborted by a consequent call
    */
   dispose() {
-    !this.disposed && this.cleanupDOM();
+    !this.disposed &&
+      this.elements.cleanupDOM({ width: this.width, height: this.height });
     this.disposed = true;
     return new Promise<boolean>((resolve, reject) => {
       const task = () => {
@@ -1654,21 +1590,6 @@ export class StaticCanvas<
         task();
       }
     });
-  }
-
-  /**
-   * Invoked as part of the **sync** operation of {@link dispose}.
-   */
-  protected cleanupDOM() {
-    const canvasElement = this.lowerCanvasEl!;
-    // restore canvas style and attributes
-    canvasElement.classList.remove('lower-canvas');
-    canvasElement.removeAttribute('data-fabric');
-    // restore canvas size to original size in case retina scaling was applied
-    canvasElement.setAttribute('width', `${this.width}`);
-    canvasElement.setAttribute('height', `${this.height}`);
-    canvasElement.style.cssText = this._originalCanvasStyle || '';
-    this._originalCanvasStyle = undefined;
   }
 
   /**
@@ -1698,10 +1619,7 @@ export class StaticCanvas<
       this.overlayImage.dispose();
     }
     this.overlayImage = null;
-    // @ts-expect-error disposing
-    this.contextContainer = null;
-    // @ts-expect-error disposing
-    this.lowerCanvasEl = undefined;
+    this.elements.dispose();
   }
 
   /**